pyproject.toml
[project]
name = "idpos-visualizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dbt-snowflake>=1.11.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "snowflake-connector-python[pandas]>=4.1.1",
]
